windowed_kmer_count_distributions: scores offsets
The offsets of the scores was previously taken as the first right
offset of a window (left = right) in order to simplify the code.
Although not wrong, it was inconsistent with how the window positions
were specified. The scores (giving the number of k-mers in each window)
should now be offset to the first base of the window.