Discrete k-mer and discrete window counts
Implemented counts of discrete k-mers or discrete windows
(non-overlapping). This was done as part of an attempt to
find a way to calculate the expected distribution of dimer counts for a
given set of nucleotide frequencies. A solution to this problem has been
implemented (very messily) in test.R (kp3.mod()). This should be moved
to kmer_spans.R and cleaned up.